refactor(macro): remove poprako-macro crate; narrow pagination and version types
- poprako-macro/: remove entire crate (AGENTS.md, Cargo.toml, src/lib.rs), along with its Cargo workspace membership and Dockerfile COPY step
- model/announcement.rs, chapter.rs, comic.rs, comment.rs, member_invitation.rs: drop #[Paginate] derive, add explicit offset/limit fields
- data/*.rs: drop #[Paginate] derives and related imports; add explicit offset/limit fields to all list-Data structs
- part/repo/step/*.rs: drop #[Paginate] derives; inline offset/limit fields into each list step struct
- All offset/limit fields: change type from u64 to u32 across model, data, handler, part, part_impl, complex, usecase, poprako-util, and tests
- All version counters (cover_version, image_version, avatar_version): change type from i64 to u32 across model, data, complex, step, usecase, RDB impl, entity, mock impl, prom, and tests
- src/part_impl/shared/result.rs: add version() helper for safe i64→u32 conversion, used in reserve_cover, reserve_image, reserve_avatar RDB impls
- rdb_impl/entity/*.rs: add #[diesel(deserialize_as = i64)] to version fields; builders accept u32 and store as i64 for Diesel
- api/http/handler/*.rs: replace #[cfg(feature = "swagger-ui")] HttpBody imports with #[allow(unused_imports)]; qualify ambiguous openapi body types (OperData→unit_data::OperData, JoinChapterData→assignment_data::JoinChapterData)
- Consolidate many use crate::data::{...} and use crate::model::{...} imports into braced groups
- .agents/skills/rust-spacing-style/check-spacing-style.py: handle attribute_item nodes in unit_anchor
- Cargo.toml: bump poprako-orchestra-extra from 0.1.0 to 0.2.0
- src/part/prom_new.rs + prom_new/: add scaffold for orchestra-based prom module
- .agents/skills/rust-spacing-style/: add test_check_spacing_style.py

+++ poprako-util/src/page.rs
@@ -3,6 +3,6 @@
 /// Offset and limit parameters for paginated database queries.
 #[derive(Debug, Clone, Copy)]
 pub struct Page {
-    pub offset: u64,
-    pub limit: u64,
+    pub offset: u32,
+    pub limit: u32,
 }
