+++ src/part_impl/repo/rdb_impl/entity/team.rs
@@ -18,7 +18,8 @@ pub struct TeamRow {
 
     pub f_avatar_key: Option<String>,
     pub f_avatar_uploaded: bool,
-    pub f_avatar_version: i64,
+    #[diesel(deserialize_as = i64)]
+    pub f_avatar_version: u32,
 
     pub f_workset_next_index: i32,
 
@@ -98,9 +99,9 @@ impl<'a> TeamAspect<'a> {
         self
     }
 
-    pub fn avatar_version(mut self, val: i64) -> Self {
+    pub fn avatar_version(mut self, val: u32) -> Self {
         //
-        self.f_avatar_version = Some(val);
+        self.f_avatar_version = Some(i64::from(val));
 
         self
     }
